Move sync out of BaseVisionManager
