+++ MapboxVisionTests/VisionManagerTests.swift
@@ -8,6 +8,9 @@ class VisionManagerTests: XCTestCase {
     var synchronizer: MockSynchronizable!
     var syncDelegate: ClosureSyncDelegate!
 
+    let otherDataSource = SyncRecordDataSource(region: .other)
+    let chinaDataSource = SyncRecordDataSource(region: .china)
+
     override func setUp() {
         super.setUp()
 
@@ -235,6 +238,33 @@ class VisionManagerTests: XCTestCase {
             SessionRecorder's action log: \(recorder.actionsLog) doesn't match expected one: \(expectedActions)
             """
         )
+    }
+
+    // MARK: - Syncing
+
+    func testChangingCountries() {
+        // Given
+        let actions: [(Country, [MockSynchronizable.Action])] = [
+            (.unknown, [.stopSync]),
+            (.USA, [
+                .stopSync,
+                .set(dataSource: otherDataSource, baseURL: URL(string: Constants.URL.defaultEventsEndpoint)!),
+                .sync
+            ]),
+            (.UK, []),
+            (.china, [
+                .stopSync,
+                .set(dataSource: chinaDataSource, baseURL: URL(string: Constants.URL.chinaEventsEndpoint)!),
+                .sync
+            ]),
+            (.unknown, [.stopSync]),
+        ]
+
+        // When
+        actions.compactMap { $0.0 }.forEach(visionManager.onCountryUpdated)
+
+        // Then
+        XCTAssertEqual(synchronizer.actionLog, Array(actions.compactMap { $0.1 }.joined()))
     }
 
     func testStopRecordingNotTriggerSyncForDefaultCountry() {
